Update storage-redundancy.md
'With GRS or GZRS, the data in the secondary region isn't available for read or write access unless there is a failover to the primary region'. 
Earlier it was 'With GRS or GZRS, the data in the secondary region isn't available for read or write access unless there is a failover to the secondary region'.